Implementation with knowledge-areas.
An "area" is a sort of "compartmentalisation" of knowledge-storage. The system uses "knowledge" for operation - here, the knowledge is selected from a list of overlapping "knowledge-areas". Operations based on the selected knowledge are exactly like on "normal" knowledge. - Just, by using multiple "knowledge areas", A LOT MORE information can be actually stored. - This system also employs "hyper-triangles" (of (only) one level up).